10 to 20 player modes for: Rev, Wiz, and Cult
Revolution, Wizard, and Cult modes now can start with only 10 players.
They adjust the difficulty by checking the player count and changing
values in the mode's pre_setup() proc or the equip_wizard() proc in the
case of wizard mode. Required number of players/enemies and other
settings were lowered to the lowest setting as default to ensure only
the lowest number of required players and enemies determines if the
round will start.

Revolution Mode:
- If there are less than 20 players, reduced recommended enemies and max
rev heads to 2 from 3
- If there are greater or equal to 20 players the recommended enemies
and max rev heads are set back to the original 3

Wizard Mode:
- in equip_wizard proc I lowered the number of uses/max uses of the
spell book to 3 and uses of the teleport scroll to 2 if there are less
than 20 players.

Cult Mode:
- for less than 20 players: required enemies = 2, recommended enemies =
4, acolytes needed for survive objective = 7, and cultists needed to
summon eldergod = 6
+ additionally: cultists required to use various runes have been
lowered: convert=2, sacrifice=2
- for 20 to less than 30 players: required enemies=4, recommended
enemies=6, acolytes needed=10, cultists required to summon the eldergod
= 9
+ additionally cultists required to use runes are their original values:
convert=3, sacrifice=3
- for 30 or more players: required enemies = 6, recommended enemies = 9,
acolytes needed=15, and runes are the same as for >= 20 players

Additionally:
- For all cult runes that check how many cultists are around the
required number of cultists has been stored as a variable so that it can
be changed
- the tome now lists the number of cultists required to use each rune,
this number is updated when the update_tome() proc is called
- the set_runecults() proc is added as a convenient way to change the
number of cultists required to activate the various runes. Each rune is
a named argument with a default value of null so you can change only the
runes specified. It also calls update_tome() at the end to keep the tome
up to date on rune requirements and as such is the preferred way to
change cultist requirements for runes.